refactor: DTO 패키지 구조를 각 계층에 맞게 분리
- 기존 `roomescape.dto` 패키지에 있던 클래스들을 `controller.dto`, `service.dto`, `repository.dto` 등 각 계층에 맞게 이동
