+++ src/main/java/roomescape/controller/AdminReservationController.java
@@ -7,8 +7,8 @@ import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
-import roomescape.dto.ReservationResult;
 import roomescape.service.ReservationService;
+import roomescape.service.dto.result.ReservationResult;
 
 import java.util.List;
