+++ src/main/java/roomescape/controller/ReservationController.java
@@ -12,15 +12,15 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
-import roomescape.dto.AvailableDateResult;
-import roomescape.dto.ReservationCreateCommand;
-import roomescape.dto.ReservationDeleteCommand;
-import roomescape.dto.ReservationDeleteRequest;
-import roomescape.dto.ReservationModifyCommand;
-import roomescape.dto.ReservationModifyRequest;
-import roomescape.dto.ReservationResult;
-import roomescape.dto.ReservationTimeStatusResult;
+import roomescape.controller.dto.ReservationDeleteRequest;
+import roomescape.controller.dto.ReservationModifyRequest;
 import roomescape.service.ReservationService;
+import roomescape.service.dto.command.ReservationCreateCommand;
+import roomescape.service.dto.command.ReservationDeleteCommand;
+import roomescape.service.dto.command.ReservationModifyCommand;
+import roomescape.service.dto.result.AvailableDateResult;
+import roomescape.service.dto.result.ReservationResult;
+import roomescape.service.dto.result.ReservationTimeStatusResult;
 
 import java.net.URI;
 import java.time.LocalDate;
