+++ src/main/java/roomescape/controller/ThemeController.java
@@ -10,9 +10,9 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
-import roomescape.dto.ThemeCreateCommand;
-import roomescape.dto.ThemeResult;
 import roomescape.service.ThemeService;
+import roomescape.service.dto.command.ThemeCreateCommand;
+import roomescape.service.dto.result.ThemeResult;
 
 import java.net.URI;
 import java.util.List;
