+++ src/main/java/roomescape/controller/dto/ReservationDeleteRequest.java
@@ -1,12 +1,9 @@
-package roomescape.dto;
+package roomescape.controller.dto;
 
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
 
-public record ReservationDeleteCommand(
-        @NotNull(message = "예약 ID는 비워둘 수 없습니다.")
-        Long reservationId,
-
+public record ReservationDeleteRequest(
         @NotNull(message = "예약자 이름은 비워둘 수 없습니다.")
         @NotBlank(message = "예약자 이름은 비워둘 수 없습니다.")
         String name
