+++ src/main/java/roomescape/service/ReservationService.java
@@ -7,19 +7,19 @@ import org.springframework.transaction.annotation.Transactional;
 import roomescape.domain.Reservation;
 import roomescape.domain.ReservationTime;
 import roomescape.domain.Theme;
-import roomescape.dto.AvailableDateResult;
-import roomescape.dto.ReservationAvailableEvent;
-import roomescape.dto.ReservationCreateCommand;
-import roomescape.dto.ReservationDeleteCommand;
-import roomescape.dto.ReservationModifyCommand;
-import roomescape.dto.ReservationResult;
-import roomescape.dto.ReservationTimeStatusResult;
 import roomescape.exception.BusinessException;
 import roomescape.exception.ErrorCode;
 import roomescape.repository.ReservationRepository;
 import roomescape.repository.ReservationTimeRepository;
 import roomescape.repository.ThemeRepository;
 import roomescape.repository.WaitingListRepository;
+import roomescape.service.dto.ReservationAvailableEvent;
+import roomescape.service.dto.command.ReservationCreateCommand;
+import roomescape.service.dto.command.ReservationDeleteCommand;
+import roomescape.service.dto.command.ReservationModifyCommand;
+import roomescape.service.dto.result.AvailableDateResult;
+import roomescape.service.dto.result.ReservationResult;
+import roomescape.service.dto.result.ReservationTimeStatusResult;
 
 import java.time.LocalDate;
 import java.time.LocalTime;
