+++ src/main/java/roomescape/service/dto/ReservationStatus.java
@@ -1,4 +1,4 @@
-package roomescape.dto;
+package roomescape.service.dto;
 
 public enum ReservationStatus {
     RESERVATION,
