+++ src/main/java/roomescape/service/dto/command/ReservationCreateCommand.java
@@ -1,4 +1,4 @@
-package roomescape.dto;
+package roomescape.service.dto.command;
 
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
@@ -6,17 +6,19 @@ import org.springframework.format.annotation.DateTimeFormat;
 
 import java.time.LocalDate;
 
-public record ReservationModifyCommand(
-        @NotNull(message = "수정할 예약 ID는 비워둘 수 없습니다.")
-        Long reservationId,
-
+public record ReservationCreateCommand(
         @NotNull(message = "예약자 이름은 비워둘 수 없습니다.")
         @NotBlank(message = "예약자 이름은 비워둘 수 없습니다.")
         String name,
 
         @DateTimeFormat(pattern = "yyyy-MM-dd")
+        @NotNull(message = "예약 날짜는 비워둘 수 없습니다.")
         LocalDate date,
 
-        Long timeId
+        @NotNull(message = "예약 시간 ID는 비워둘 수 없습니다.")
+        Long timeId,
+
+        @NotNull(message = "예약 테마 ID는 비워둘 수 없습니다.")
+        Long themeId
 ) {
 }
