+++ src/main/java/roomescape/service/dto/command/ReservationModifyCommand.java
@@ -1,4 +1,4 @@
-package roomescape.dto;
+package roomescape.service.dto.command;
 
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
@@ -6,7 +6,10 @@ import org.springframework.format.annotation.DateTimeFormat;
 
 import java.time.LocalDate;
 
-public record ReservationModifyRequest(
+public record ReservationModifyCommand(
+        @NotNull(message = "수정할 예약 ID는 비워둘 수 없습니다.")
+        Long reservationId,
+
         @NotNull(message = "예약자 이름은 비워둘 수 없습니다.")
         @NotBlank(message = "예약자 이름은 비워둘 수 없습니다.")
         String name,
