+++ src/main/java/roomescape/service/dto/command/ReservationTimeCreateCommand.java
@@ -1,4 +1,4 @@
-package roomescape.dto;
+package roomescape.service.dto.command;
 
 import jakarta.validation.constraints.NotNull;
 import org.springframework.format.annotation.DateTimeFormat;
