+++ src/main/java/roomescape/service/dto/result/AvailableDateResult.java
@@ -1,4 +1,4 @@
-package roomescape.dto;
+package roomescape.service.dto.result;
 
 import java.time.LocalDate;
 import java.util.List;
