+++ src/test/java/roomescape/controller/ReservationControllerTest.java
@@ -5,9 +5,9 @@ import io.restassured.http.ContentType;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.jdbc.core.JdbcTemplate;
-import roomescape.dto.AvailableDateResult;
-import roomescape.dto.ReservationResult;
-import roomescape.dto.ReservationTimeStatusResult;
+import roomescape.service.dto.result.AvailableDateResult;
+import roomescape.service.dto.result.ReservationResult;
+import roomescape.service.dto.result.ReservationTimeStatusResult;
 import roomescape.support.SpringBootApiTest;
 
 import java.time.LocalDate;
