+++ src/test/java/roomescape/service/ReservationServiceTest.java
@@ -10,20 +10,20 @@ import org.springframework.context.ApplicationEventPublisher;
 import roomescape.domain.Reservation;
 import roomescape.domain.ReservationTime;
 import roomescape.domain.Theme;
-import roomescape.dto.ReservationAvailableEvent;
-import roomescape.dto.ReservationCreateCommand;
-import roomescape.dto.ReservationDeleteCommand;
-import roomescape.dto.ReservationModifyCommand;
-import roomescape.dto.ReservationResult;
-import roomescape.dto.ReservationStatus;
-import roomescape.dto.ReservationTimeStatusResult;
-import roomescape.dto.ReservationTimesWithStatus;
 import roomescape.exception.BusinessException;
 import roomescape.exception.ErrorCode;
 import roomescape.repository.ReservationRepository;
 import roomescape.repository.ReservationTimeRepository;
 import roomescape.repository.ThemeRepository;
 import roomescape.repository.WaitingListRepository;
+import roomescape.repository.dto.ReservationTimesWithStatus;
+import roomescape.service.dto.ReservationAvailableEvent;
+import roomescape.service.dto.ReservationStatus;
+import roomescape.service.dto.command.ReservationCreateCommand;
+import roomescape.service.dto.command.ReservationDeleteCommand;
+import roomescape.service.dto.command.ReservationModifyCommand;
+import roomescape.service.dto.result.ReservationResult;
+import roomescape.service.dto.result.ReservationTimeStatusResult;
 
 import java.time.LocalDate;
 import java.time.LocalTime;
