+++ src/test/java/roomescape/service/ReservationTimeServiceTest.java
@@ -6,13 +6,13 @@ import org.mockito.InjectMocks;
 import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
 import roomescape.domain.ReservationTime;
-import roomescape.dto.ReservationTimeCreateCommand;
-import roomescape.dto.ReservationTimeResult;
 import roomescape.exception.BusinessException;
 import roomescape.exception.ErrorCode;
 import roomescape.repository.ReservationRepository;
 import roomescape.repository.ReservationTimeRepository;
 import roomescape.repository.WaitingListRepository;
+import roomescape.service.dto.command.ReservationTimeCreateCommand;
+import roomescape.service.dto.result.ReservationTimeResult;
 
 import java.time.LocalTime;
 import java.util.List;
